% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grasp.R
\name{fgrasp_statistic}
\alias{fgrasp_statistic}
\title{GRASP test statistics}
\usage{
fgrasp_statistic(y, model, L = 50)
}
\arguments{
\item{y}{n labels $y_j \in \{0, 1\}$ as a vector}

\item{model}{the model $\hat{\eta}:\chi \to \link{0,1}$ as a vector of predicted labels, the result of running the model on each x_j.}

\item{L}{an integer greater than or equal to 1, default to 50}
}
\description{
This function will generate the GRASP test statistics defined as $V_{n,L}$ in Algorithm 2 from the original paper, in the distribution free setting with score function T(x,w) = w
}
