% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRASP.R
\name{GRASP_L}
\alias{GRASP_L}
\title{Estimate the required sample size based on AUC for Study 1}
\usage{
GRASP_L(theta, h, alpha = 0.05, L = 0.1)
}
\arguments{
\item{theta}{The guess of the AUC value that the prediction model can potential achieve.}

\item{h}{The ratio of non-diseased to diseased patients, and we always assume that h >= 1.}

\item{alpha}{The significance level used to compute the (1-alpha)100\% CI. Default to 0.05,
which results in a 95\% CI.}

\item{L}{The threshold of the length of the (1-alpha)100\% CI. Default to 0.1.}
}
\value{
\code{sampsize} A matrix summarizing the lower and upper bounds of the required sample size
for the diseased (n1), non-diseased (n0), and the entire (n) patient group.
}
\description{
This function calculates the lower and upper bounds of the required sample size
for the prediction model development to ensure that the length of the (1-alpha)100\%
confidence interval (CI) for the empirical AUC estimator remains within a specific threshold.
}
