% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRASP.R
\name{GRASP_test}
\alias{GRASP_test}
\title{Estimate the required sample size based on AUC for Study 2}
\usage{
GRASP_test(theta0, theta1, h, alpha = 0.05, beta = 0.2)
}
\arguments{
\item{theta0}{The AUC value specified in the null hypothesis, H0: AUC=theta0.}

\item{theta1}{The AUC value specified in the alternative hypothesis, Ha: AUC=theta1>thata0.}

\item{h}{The ratio of non-diseased to diseased patients, and we always assume that h >= 1.}

\item{alpha}{The desired type I error control. Default to 0.05.}

\item{beta}{The desired type II error control. Default to 0.2, which results in
the level of statistical power to be 0.8.}
}
\value{
\code{sampsize} A matrix summarizing the lower and upper bounds of the required sample size
for the diseased (n1), non-diseased (n0), and the entire (n) patient group.
}
\description{
This function calculates the lower and upper bounds of the required sample size
for the prediction model development to test hypothesis with a desired type I error
control and a predetermined level of statistical power.
}
