% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grasp.R
\name{hypothesis_tv}
\alias{hypothesis_tv}
\title{Total Variation Distance test statistic}
\usage{
hypothesis_tv(eta, eta_hat)
}
\arguments{
\item{eta}{the vector of labels in the test data}

\item{eta_hat}{the vector of labels predicted by the model}
}
\description{
This function will return the initial hypothesis test statistic generated by equation(3) in the paper, with $f(t) = 1/2|t-1|$.
}
