% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRASP_simulation.R
\name{sampsize_simulation}
\alias{sampsize_simulation}
\title{Perform simulation to compare different sample size determination methods}
\usage{
sampsize_simulation(
  N,
  d,
  cov_mat,
  beta0,
  coeff,
  X_index,
  diff_AUC = 0.05,
  alpha = 0.05,
  beta = 0.2,
  L = 0.1,
  seed,
  nBoots = 500,
  nCores = 1,
  model = c("LR", "LDA", "naive_bayes", "svmLinear")
)
}
\arguments{
\item{N}{The sample size of the underlying whole target population. Should be a
large number (e.g., N=100,000).}

\item{d}{The dimension of the feature vector X.}

\item{cov_mat}{The covariance matrix of the features. If missing, feature vector X
will be generated independently and identically from a normal distribution with mean 0
and standard deviation 1. If specified, feature vector X will be generated by a multivariate
normal distribution with mean 0 and covariance matrix cov_mat.}

\item{beta0}{The coefficient value for the intercept of the underlying model that determines
the probability of each patient developing the disease status.}

\item{coeff}{Vector of the coefficient values for the feature vector X of the underlying model that determines
the probability of each patient developing the disease status. Dimension of coeff should be equal to d.}

\item{X_index}{The indices of the features assumed to be included in the prediction model.}

\item{diff_AUC}{The difference of the AUC value in the hypothesis test (i.e., theta1-theta0).
Default to 0.05.}

\item{alpha}{The significance level used to compute the (1-alpha)100\% CI in Study 1 and the
desired type I error control in Study 2. Default to 0.05.}

\item{beta}{The desired type II error control in Study 2. Default to 0.2, which results in
the level of statistical power to be 0.8.}

\item{L}{The threshold of the length of the (1-alpha)100\% CI. Default to 0.1.}

\item{seed}{The random seed value.}

\item{nBoots}{The number of bootstraps. Default to 500.}

\item{nCores}{Number of cores to use in computing results. Set to 1 to not use parallel computing.
Defaults to 1.}

\item{model}{Type of the prediction model. Can take value as "LR" (logistic regression), "LDA" (linear
distriminant analysis), "naive_bayes", and "svmLinear" (SVM with linear kernel). Other machine learning
(ML) algorithms incorporated in R package 'caret' can also be specified.}
}
\value{
\code{data_n} A matrix summarizing the mean and sd of the estimated sample size for
the diseased (n1), non-diseased (n0), and the entire (n) patient group based on different SSD methods.

\code{data_L} A matrix summarizing the mean adn sd of the CI length based on different SSD methods.

\code{data_power} A matrix summarizing the power of hypothesis test based on different SSD methods.

\code{data_var} A matrix summarizing the true positive rate (TPR) and false positive rate (FPR) in
identifying the significant and insignificant variables in the prediction model. Only returned if
logistic regression model is fitted.
}
\description{
This function generates pseudo-datasets to assess the performance of different
sample size determination (SSD) methods for estimating the required sample size for
prediction model development in both Study 1 and Study 2. It can compare the
effectiveness of models such as logistic regression (LR), linear discriminant
analysis (LDA), naive Bayes (NB), and support vector machine (SVM) with a linear
kernel, against the rules-of-thumb and Riley's method.
}
